requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.21.0
opencv-python>=4.5.0
kaggle>=1.5.0
requests>=2.28.0
opendatasets>=0.1.22